requirements.txt
asgiref==3.8.1
Django==5.2
mysql-connector==2.2.9
mysqlclient==2.2.7
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
pip==25.0.1
bcrypt==4.1.2
openpyxl==3.1.5
reportlab==3.6.8
tabulate==0.9.0